removed mariadb npm dependency - read description
due to using the filesystem dependency to issue SQL commands from the cli, this library isn't usable in the browser. We will replace it with the typeorm npm library.

Also:
removed functions that depend on mariadb library
corrected the interface for IIMPDatabaseTable, hopefully?
renamed env variables to be usable by react-scripts
added explicit NODE_ENV declarations in scripts
changed mockState to createEmptyContext for db